Created newsletterform component
